delete Event by Id

+++ src/main/java/com/tiendat/spring_webmvc/BootDemo/respository/EventRepository.java
@@ -28,7 +28,7 @@ public interface EventRepository extends JpaRepository<Event, Long>{
 	
 	Event save(Event event);
 	Event saveAndFlush(Event event);
-	int deleteByEventId(int eventId);
+	Long deleteByEventId(int eventId);
 	
 	
 }
